requirements.txt
django-4.0
djangorestframework
django-rest-swagger
pyYaml
gunicorn
psycopg2